deck level calculation

+++ app/src/main/java/ml/nandor/confusegroups/domain/model/Review.kt
@@ -20,5 +20,5 @@ data class Review(
 
     // unix timestamp
     @PrimaryKey
-    val timeStamp: Int
+    val timeStamp: Long
 )
